Debug console window and assorted fixes

+++ ssdv.c
@@ -410,7 +410,7 @@ static char ssdv_out_jpeg_int(ssdv_t *s, uint8_t rle, int value)
 	jpeg_encode_int(value, &intbits, &intlen);
 	r = jpeg_dht_lookup_symbol(s, (rle << 4) | (intlen & 0x0F), &huffbits, &hufflen);
 	
-	if(r != SSDV_OK) fprintf(stderr, "jpeg_dht_lookup_symbol: %i (%i:%i)\n", r, value, rle);
+    if(r != SSDV_OK) fprintf(stderr, "jpeg_dht_lookup_symbol: %i (%i:%i)\n", r, value, rle);
 	
 	ssdv_outbits(s, huffbits, hufflen);
 	if(intlen) ssdv_outbits(s, intbits, intlen);
@@ -666,7 +666,7 @@ static char ssdv_have_marker(ssdv_t *s)
 	
 	case J_SOF2:
 		/* Don't do progressive images! */
-		fprintf(stderr, "Error: Progressive images not supported\n");
+        fprintf(stderr, "Error: Progressive images not supported\n");
 		return(SSDV_ERROR);
 	
 	case J_EOI:
@@ -711,35 +711,35 @@ static char ssdv_have_marker_data(ssdv_t *s)
 		s->height = (d[1] << 8) | d[2];
 		
 		/* Display information about the image... */
-		fprintf(stderr, "Precision: %i\n", d[0]);
-		fprintf(stderr, "Resolution: %ix%i\n", s->width, s->height);
-		fprintf(stderr, "Components: %i\n", d[5]);
+        fprintf(stderr, "Precision: %i\n", d[0]);
+        fprintf(stderr, "Resolution: %ix%i\n", s->width, s->height);
+        fprintf(stderr, "Components: %i\n", d[5]);
 		
 		/* The image must have a precision of 8 */
 		if(d[0] != 8)
 		{
-			fprintf(stderr, "Error: The image must have a precision of 8\n");
+            fprintf(stderr, "Error: The image must have a precision of 8\n");
 			return(SSDV_ERROR);
 		}
 		
 		/* The image must have 3 components (Y'Cb'Cr) */
 		if(d[5] != 3)
 		{
-			fprintf(stderr, "Error: The image must have 3 components\n");
+            fprintf(stderr, "Error: The image must have 3 components\n");
 			return(SSDV_ERROR);
 		}
 		
 		/* Maximum image is 4080x4080 */
 		if(s->width > 4080 || s->height > 4080)
 		{
-			fprintf(stderr, "Error: The image is too big. Maximum resolution is 4080x4080\n");
+            fprintf(stderr, "Error: The image is too big. Maximum resolution is 4080x4080\n");
 			return(SSDV_ERROR);
 		}
 		
 		/* The image dimensions must be a multiple of 16 */
 		if((s->width & 0x0F) || (s->height & 0x0F))
 		{
-			fprintf(stderr, "Error: The image dimensions must be a multiple of 16\n");
+            fprintf(stderr, "Error: The image dimensions must be a multiple of 16\n");
 			return(SSDV_ERROR);
 		}
 		
@@ -750,11 +750,11 @@ static char ssdv_have_marker_data(ssdv_t *s)
 			uint8_t *dq = &d[i * 3 + 6];
 			if(dq[0] != i + 1)
 			{
-				fprintf(stderr, "Error: Components are not in order in the SOF0 header\n");
+                fprintf(stderr, "Error: Components are not in order in the SOF0 header\n");
 				return(SSDV_ERROR);
 			}
 			
-			fprintf(stderr, "DQT table for component %i: %02X, Sampling factor: %ix%i\n", dq[0], dq[2], dq[1] & 0x0F, dq[1] >> 4);
+            fprintf(stderr, "DQT table for component %i: %02X, Sampling factor: %ix%i\n", dq[0], dq[2], dq[1] & 0x0F, dq[1] >> 4);
 			
 			/* The first (Y) component must have a factor of 2x2,2x1,1x2 or 1x1 */
 			if(dq[0] == 1)
@@ -766,13 +766,13 @@ static char ssdv_have_marker_data(ssdv_t *s)
 				case 0x21: s->mcu_mode = 2; s->ycparts = 2; break;
 				case 0x11: s->mcu_mode = 3; s->ycparts = 1; break;
 				default:
-					fprintf(stderr, "Error: Component 1 sampling factor is not supported\n");
+                    fprintf(stderr, "Error: Component 1 sampling factor is not supported\n");
 					return(SSDV_ERROR);
 				}
 			}
 			else if(dq[0] != 1 && dq[1] != 0x11)
 			{
-				fprintf(stderr, "Error: Component %i sampling factor must be 1x1\n", dq[0]);
+                fprintf(stderr, "Error: Component %i sampling factor must be 1x1\n", dq[0]);
 				return(SSDV_ERROR);
 			}
 		}
@@ -786,11 +786,11 @@ static char ssdv_have_marker_data(ssdv_t *s)
 		case 3: l = (s->width >> 3) * (s->height >> 3); break;
 		}
 		
-		fprintf(stderr, "MCU blocks: %i\n", (int) l);
+        fprintf(stderr, "MCU blocks: %i\n", (int) l);
 		
 		if(l > 0xFFFF)
 		{
-			fprintf(stderr, "Error: Maximum number of MCU blocks is 65535\n");
+            fprintf(stderr, "Error: Maximum number of MCU blocks is 65535\n");
 			return(SSDV_ERROR);
 		}
 		
@@ -799,12 +799,12 @@ static char ssdv_have_marker_data(ssdv_t *s)
 		break;
 	
 	case J_SOS:
-		fprintf(stderr, "Components: %i\n", d[0]);
+        fprintf(stderr, "Components: %i\n", d[0]);
 		
 		/* The image must have 3 components (Y'Cb'Cr) */
 		if(d[0] != 3)
 		{
-			fprintf(stderr, "Error: The image must have 3 components\n");
+            fprintf(stderr, "Error: The image must have 3 components\n");
 			return(SSDV_ERROR);
 		}
 		
@@ -813,11 +813,11 @@ static char ssdv_have_marker_data(ssdv_t *s)
 			uint8_t *dh = &d[i * 2 + 1];
 			if(dh[0] != i + 1)
 			{
-				fprintf(stderr, "Error: Components are not in order in the SOF0 header\n");
+                fprintf(stderr, "Error: Components are not in order in the SOF0 header\n");
 				return(SSDV_ERROR);
 			}
 			
-			fprintf(stderr, "Component %i DHT: %02X\n", dh[0], dh[1]);
+            fprintf(stderr, "Component %i DHT: %02X\n", dh[0], dh[1]);
 		}
 		
 		/* Do I need to look at the last three bytes of the SOS data? */
@@ -826,14 +826,14 @@ static char ssdv_have_marker_data(ssdv_t *s)
 		/* Verify all of the DQT and DHT tables where loaded */
 		if(!s->sdqt[0] || !s->sdqt[1])
 		{
-			fprintf(stderr, "Error: The image is missing one or more DQT tables\n");
+            fprintf(stderr, "Error: The image is missing one or more DQT tables\n");
 			return(SSDV_ERROR);
 		}
 		
 		if(!s->sdht[0][0] || !s->sdht[0][1] ||
 		   !s->sdht[1][0] || !s->sdht[1][1])
 		{
-			fprintf(stderr, "Error: The image is missing one or more DHT tables\n");
+            fprintf(stderr, "Error: The image is missing one or more DHT tables\n");
 			return(SSDV_ERROR);
 		}
 		
@@ -883,7 +883,7 @@ static char ssdv_have_marker_data(ssdv_t *s)
 	
 	case J_DRI:
 		s->dri = (d[0] << 8) + d[1];
-		fprintf(stderr, "Reset interval: %i blocks\n", s->dri);
+        fprintf(stderr, "Reset interval: %i blocks\n", s->dri);
 		break;
 	}
 	
@@ -1070,7 +1070,7 @@ char ssdv_enc_get_packet(ssdv_t *s)
 			else if(r != SSDV_FEED_ME)
 			{
 				/* An error occured */
-				fprintf(stderr, "ssdv_process() failed: %i\n", r);
+                fprintf(stderr, "ssdv_process() failed: %i\n", r);
 				return(SSDV_ERROR);
 			}
 			break;
@@ -1265,12 +1265,12 @@ char ssdv_dec_feed(ssdv_t *s, uint8_t *packet)
 		}
 		
 		/* Display information about the image */
-		fprintf(stderr, "Callsign: %s\n", decode_callsign(callsign, s->callsign));
-		fprintf(stderr, "Image ID: %02X\n", s->image_id);
-		fprintf(stderr, "Resolution: %ix%i\n", s->width, s->height);
-		fprintf(stderr, "MCU blocks: %i\n", s->mcu_count);
-		fprintf(stderr, "Sampling factor: %s\n", factor);
-		fprintf(stderr, "Quality level: %d\n", s->quality);
+        fprintf(stderr, "Callsign: %s\n", decode_callsign(callsign, s->callsign));
+        fprintf(stderr, "Image ID: %02X\n", s->image_id);
+        fprintf(stderr, "Resolution: %ix%i\n", s->width, s->height);
+        fprintf(stderr, "MCU blocks: %i\n", s->mcu_count);
+        fprintf(stderr, "Sampling factor: %s\n", factor);
+        fprintf(stderr, "Quality level: %d\n", s->quality);
 		
 		/* Output JPEG headers and enable byte stuffing */
 		ssdv_out_headers(s);
@@ -1281,7 +1281,7 @@ char ssdv_dec_feed(ssdv_t *s, uint8_t *packet)
 	if(packet_id != s->packet_id)
 	{
 		/* One or more packets have been lost! */
-		fprintf(stderr, "Gap detected between packets %i and %i\n", s->packet_id - 1, packet_id);
+        fprintf(stderr, "Gap detected between packets %i and %i\n", s->packet_id - 1, packet_id);
 		
 		/* If this packet has no new MCU, ignore */
 		if(s->packet_mcu_offset == 0xFF) return(SSDV_FEED_ME);
@@ -1329,7 +1329,7 @@ char ssdv_dec_feed(ssdv_t *s, uint8_t *packet)
 		else if(r != SSDV_FEED_ME)
 		{
 			/* An error occured */
-			fprintf(stderr, "ssdv_process() failed: %i\n", r);
+            fprintf(stderr, "ssdv_process() failed: %i\n", r);
 			return(SSDV_ERROR);
 		}
 	}
@@ -1385,7 +1385,7 @@ char ssdv_dec_is_packet(uint8_t *packet, int *errors)
 		x = crc32(&pkt[1], pkt_size_crcdata);
 		
 		i = 1 + pkt_size_crcdata;
-		if(x == (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
+        if(x == (uint32_t) (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
 		{
 			/* Valid, set the type and continue */
 			type = SSDV_TYPE_NOFEC;
@@ -1404,7 +1404,7 @@ char ssdv_dec_is_packet(uint8_t *packet, int *errors)
 		x = crc32(&pkt[1], pkt_size_crcdata);
 		
 		i = 1 + pkt_size_crcdata;
-		if(x == (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
+        if(x == (uint32_t) (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
 		{
 			/* Valid, set the type and continue */
 			type = SSDV_TYPE_NORMAL;
@@ -1428,7 +1428,7 @@ char ssdv_dec_is_packet(uint8_t *packet, int *errors)
 		x = crc32(&pkt[1], pkt_size_crcdata);
 		
 		i = 1 + pkt_size_crcdata;
-		if(x == (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
+        if(x == (uint32_t) (pkt[i + 3] | (pkt[i + 2] << 8) | (pkt[i + 1] << 16) | (pkt[i] << 24)))
 		{
 			/* Valid, set the type and continue */
 			type = SSDV_TYPE_NORMAL;
